Move fixtures to a pytest plugin
[noissue]
